forgot to change the dtype of the new one hot embedding

+++ src/environments/fine_alignment/all_dict_obs_env_diff_img_diff_rew_one_hot_id.py
@@ -48,6 +48,6 @@ class AllDictObsEnvDiffImgDiffRewOneHotID(AllDictObsEnvDiffImgDiffRew):
     
     def get_observation(self):
         observation = super().get_observation()
-        observation["current_panel_id"] = np.zeros(self.telescope.n_panels)
+        observation["current_panel_id"] = np.zeros(self.telescope.n_panels, dtype=np.uint8)
         observation["current_panel_id"][self.current_panel] = 1
         return observation
